Touched up turret PID, remove threaded pose estimation, reset pose on construct
