+++ .gitignore
@@ -1,3 +1,7 @@
+# silly intellij/android studio things
+.idea/
+local.properties
+
 # This gitignore has been specially created by the WPILib team.
 # If you remove items from this file, intellisense might break.
